chore: split manifest schema model
